feat: handle enter chatroom
- 處理 重複進入聊天室例外

+++ app/src/main/java/com/example/os_finalproject/Data/API.kt
@@ -4,5 +4,5 @@ package com.example.os_finalproject.Data
 data class RoomInfoRes(val result: Result) {
     data class Result(val roomInfoList: ArrayList<RoomInfoList>)
     data class RoomInfoList(val roomID: String, val roomUser: ArrayList<RoomUser>)
-    data class RoomUser(val userName: String, val socketID: String, val refreshTime: Int)
+    data class RoomUser(val userName: String, val socketID: String, val refreshTime: Long)
 }
